chore: header refinements
