+++ src/App/App.jsx
@@ -5,7 +5,7 @@ function App() {
   return (
     <div className="App">
       <Header />
-      {/* <MyRouter /> */}
+      <MyRouter />
     </div>
   );
 }
