+++ src/Components/About/About.jsx
@@ -1,20 +1,14 @@
 import React from 'react';
-import { Link } from "react-router-dom";
 
 const About = () => {
   return (
-    <>
-      <main>
-        <h2>Who are we?</h2>
-        <p>
-          That feels like an existential question, don't you
-          think?
-        </p>
-      </main>
-      <nav>
-        <Link to="/">Home</Link>
-      </nav>
-    </>
+    <main>
+      <h2>Who are we?</h2>
+      <p>
+        That feels like an existential question, don't you
+        think?
+      </p>
+    </main>
   );
 }
