+++ src/Components/Home/Home.jsx
@@ -1,13 +1,11 @@
 import React from 'react';
-import { Link } from "react-router-dom";
 
 const Home = () => {
   return (
-    <>
-      <nav>
-        <Link to="/about">About</Link>
-      </nav>
-    </>
+    <main>
+      <h2>Welcome to the homepage!</h2>
+      <p>You can do this, I believe in you.</p>
+    </main>
   );
 }
