+++ src/Constants/pages.js
@@ -1,0 +1,10 @@
+export const pages = [
+  {
+    displayName: 'Home',
+    path: '/'
+  },
+  {
+    displayName: 'About',
+    path: '/about'
+  }
+]
